utils to pythium, binning to its own module
