+++ misc/legacy/histogramming/Konrad/cross_product_backend.py
@@ -13,7 +13,7 @@ currentdir = os.path.dirname(os.path.realpath(__file__))
 parentdir = os.path.dirname(currentdir)
 sys.path.append(parentdir)
 
-from utils.common.samples import Sample
+from pythium.common.samples import Sample
 
 
 class XP_Sample:
