+++ pythium/common/functor.py
@@ -1,5 +1,5 @@
 
-from utils.common.tools import Evaluator
+from pythium.common.tools import Evaluator
 # ========= type hinting
 from typing import Optional, Union, Type, TypeVar
 from beartype.typing import List, Callable, Dict
