+++ pythium/common/selection.py
@@ -1,5 +1,5 @@
-from utils.common.tools import Evaluator
-from utils.common.functor import Functor
+from pythium.common.tools import Evaluator
+from pythium.common.functor import Functor
 # ========= type hinting
 from typing import Optional, Union, Type, TypeVar
 from beartype.typing import List, Callable, Dict
