+++ pythium/common/tools.py
