+++ pythium/sklimming/config.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, List, Optional, Union
 from pathlib import Path 
-from utils.common import logger 
+from pythium.common import logger 
 from pydoc import importfile
 import types
 logger = logger.ColoredLogger()
