+++ tests/histogramming/expr_evaluator_test.py
@@ -2,7 +2,7 @@ import os,sys
 currentdir = os.path.dirname(os.path.realpath(__file__))
 parentdir = os.path.dirname(currentdir)
 sys.path.append(parentdir)
-from utils.common.tools import Evaluator
+from pythium.common.tools import Evaluator
 import awkward as ak
 import numpy as np
