--- utils/histogramming/__init__.py
@@ -1,4 +1,0 @@
-
-from . import config
-from . import processor
-from . import objects
